nvd_destroy_and_quit: Destroy widgets once done with the dialog.
When we are done with a dialog, we shouldn't just call `gtk_main_quit` as this lets the previous dialog exist and even outlive newer dialogs. Instead, we should destroy the entire widget first and then quit the Gtk main loop, which gives us more fine-grained control over the dialog.