CMakeLists.txt: Remove hardcoded /usr prefix on GNU/Linux
Allow the user to choose where the library is installed instead of forcing /usr as the prefix. Distributions should also start making use of /usr/local (The default install prefix for CMake on GNU/Linux). For those interested, take a look at /etc/ld.so.conf and the relevant manpages to see how to add a new path for the linker to search. Alternatively, set CMAKE_INSTALL_PREFIX=/usr at build/configure time to get the old behaviour back.